added the popular movies
