+++ src/index.html
@@ -21,17 +21,17 @@
                             <li class="nav-item" role="presentation">
                                 <button class="nav-link active" id="upcoming-tab" data-bs-toggle="tab"
                                     data-bs-target="#upcoming" type="button" role="tab" aria-controls="upcoming"
-                                    aria-selected="true"><span class="movies-counter">4 </span>Coming Soons </button>
+                                    aria-selected="true"><span class="movies-counter">4 </span>Upcoming</button>
                             </li>
                             <li class="nav-item" role="presentation">
-                                <button class="nav-link" id="top-tab" data-bs-toggle="tab" data-bs-target="#top"
-                                    type="button" role="tab" aria-controls="top"
-                                    aria-selected="false">Top 250</button>
+                                <button class="nav-link" id="latest-tab" data-bs-toggle="tab" data-bs-target="#latest"
+                                    type="button" role="tab" aria-controls="latest"
+                                    aria-selected="false">Latest</button>
                             </li>
                             <li class="nav-item" role="presentation">
-                                <button class="nav-link" id="liked-tab" data-bs-toggle="tab" data-bs-target="#liked"
-                                    type="button" role="tab" aria-controls="liked"
-                                    aria-selected="false">Most Liked</button>
+                                <button class="nav-link" id="popular-tab" data-bs-toggle="tab" data-bs-target="#popular"
+                                    type="button" role="tab" aria-controls="popular"
+                                    aria-selected="false">Popular</button>
                             </li>
                         </ul>
                     </div>
@@ -40,9 +40,9 @@
         </nav>
     </header>
     <div class="tab-content" id="myTabContent">
-        <div class="tab-pane fade show active d-flex" id="upcoming" role="tabpanel" aria-labelledby="upcoming-tab"></div>
-        <div class="tab-pane fade" id="top" role="tabpanel" aria-labelledby="top-tab"></div>
-        <div class="tab-pane fade" id="liked" role="tabpanel" aria-labelledby="liked-tab"></div>
+        <div class="tab-pane fade show active" id="upcoming" role="tabpanel" aria-labelledby="upcoming-tab"></div>
+        <div class="tab-pane fade" id="latest" role="tabpanel" aria-labelledby="latest-tab"></div>
+        <div class="tab-pane fade" id="popular" role="tabpanel" aria-labelledby="popular-tab"></div>
     </div>
     <footer class="bg-dark text-center mt-4">
         <div class="container p-4 pb-0">
